+++ static/composite-types-simple.sql
@@ -13,3 +13,18 @@ CREATE TYPE person_address AS
     address_line_1 text,
     address_line_2 text
   );
+
+-- The below types and tables serve to show that composite types inform which scalar types are relevant.
+
+CREATE TYPE discoverable_types AS
+  (
+    only_occurring_here1 bit
+  );
+
+CREATE TABLE discoverable_types_root_occurrence
+  (
+    col discoverable_types
+  );
+
+-- We add this because ndc-test is going to check that this table it not empty.
+INSERT INTO discoverable_types_root_occurrence(col) VALUES (ROW(1)::discoverable_types);